requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
argcomplete==3.1.1
async-timeout==4.0.3
attrs==23.1.0
charset-normalizer==3.2.0
colorama==0.4.6
commitizen==3.6.0
decli==0.6.1
discord==2.3.2
discord.py==2.3.2
frozenlist==1.4.0
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-dataclass==8.5.14
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.1
prompt-toolkit==3.0.39
python-decouple==3.8
PyYAML==6.0.1
questionary==1.10.0
termcolor==2.3.0
tomlkit==0.12.1
typing-inspect==0.9.0
typing_extensions==4.7.1
wcwidth==0.2.6
wiseyoda==1.0.1
yarl==1.9.2
zipp==3.16.2
